# qontract_api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMergeRequestRequest": {
        "description": "Request to create a merge request in a VCS repository.",
        "properties": {
          "auto_merge": {
            "default": false,
            "description": "Whether to enable auto-merge",
            "title": "Auto Merge",
            "type": "boolean"
          },
          "description": {
            "default": "",
            "description": "Merge request description",
            "title": "Description",
            "type": "string"
          },
          "file_operations": {
            "description": "File operations to include in the MR",
            "items": {
              "$ref": "#/components/schemas/MergeRequestFileOperation"
            },
            "title": "File Operations",
            "type": "array"
          },
          "labels": {
            "description": "Labels to apply to the MR",
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "repo_url": {
            "description": "Repository URL (e.g., https://gitlab.com/group/project)",
            "title": "Repo Url",
            "type": "string"
          },
          "source_branch": {
            "description": "Source branch name",
            "title": "Source Branch",
            "type": "string"
          },
          "target_branch": {
            "default": "master",
            "description": "Target branch name",
            "title": "Target Branch",
            "type": "string"
          },
          "title": {
            "description": "Merge request title",
            "title": "Title",
            "type": "string"
          },
          "token": {
            "$ref": "#/components/schemas/Secret",
            "description": "Secret reference for VCS API token"
          }
        },
        "required": [
          "repo_url",
          "token",
          "title",
          "source_branch",
          "file_operations"
        ],
        "title": "CreateMergeRequestRequest",
        "type": "object"
      },
      "CreateMergeRequestResponse": {
        "description": "Response after creating a merge request.",
        "properties": {
          "url": {
            "description": "URL of the created merge request",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "CreateMergeRequestResponse",
        "type": "object"
      },
...
      "MergeRequestFileOperation": {
        "description": "A file operation within a merge request.\n\nSet ``content`` to ``None`` to delete the file.",
        "properties": {
          "commit_message": {
            "description": "Commit message for this file change",
            "title": "Commit Message",
            "type": "string"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "File content (None = delete the file)",
            "title": "Content"
          },
          "path": {
            "description": "File path in the repository",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "content",
          "commit_message"
        ],
        "title": "MergeRequestFileOperation",
        "type": "object"
      },
...
      "RepoOwnersResponse": {
        "description": "Response model for repository OWNERS file data.\n\nAttention: usernames are provider-specific (e.g., GitHub usernames).",
        "properties": {
          "approvers": {
            "description": "List of usernames who can approve changes",
            "items": {
              "type": "string"
            },
            "title": "Approvers",
            "type": "array"
          },
          "provider": {
            "$ref": "#/components/schemas/VCSProvider",
            "description": "VCS provider type"
          },
          "reviewers": {
            "description": "List of usernames who can review changes",
            "items": {
              "type": "string"
            },
            "title": "Reviewers",
            "type": "array"
          }
        },
        "required": [
          "provider"
        ],
        "title": "RepoOwnersResponse",
        "type": "object"
      },
...
      "Secret": {
        "description": "Reference to a secret stored in a secret manager.",
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Specific field within the secret",
            "title": "Field"
          },
          "path": {
            "description": "Path to the secret",
            "title": "Path",
            "type": "string"
          },
          "secret_manager_url": {
            "description": "Secret Manager URL",
            "title": "Secret Manager Url",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Version of the secret",
            "title": "Version"
          }
        },
        "required": [
          "secret_manager_url",
          "path"
        ],
        "title": "Secret",
        "type": "object"
      },
...
      "VCSProvider": {
        "description": "VCS provider types.\n\nExtensible enum for supported VCS providers.",
        "enum": [
          "github",
          "gitlab"
        ],
        "title": "VCSProvider",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for qontract-reconcile api integrations",
    "summary": "Qontract API OpenAPI schema",
    "title": "Qontract API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/external/vcs/merge-requests": {
      "get": {
        "description": "Find an open merge request by source branch.\n\nArgs:\n    params: Query parameters with repo_url, source_branch, and token\n\nReturns:\n    CreateMergeRequestResponse with the MR URL, or None if not found",
        "operationId": "vcs-find-merge-request",
        "parameters": [
          {
            "description": "Secret Manager URL",
            "in": "query",
            "name": "secret_manager_url",
            "required": true,
            "schema": {
              "description": "Secret Manager URL",
              "title": "Secret Manager Url",
              "type": "string"
            }
          },
          {
            "description": "Path to the secret",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "description": "Path to the secret",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "description": "Specific field within the secret",
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specific field within the secret",
              "title": "Field"
            }
          },
          {
            "description": "Version of the secret",
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Version of the secret",
              "title": "Version"
            }
          },
          {
            "description": "Repository URL (e.g., https://gitlab.com/group/project)",
            "in": "query",
            "name": "repo_url",
            "required": true,
            "schema": {
              "description": "Repository URL (e.g., https://gitlab.com/group/project)",
              "title": "Repo Url",
              "type": "string"
            }
          },
          {
            "description": "Source branch name to search for",
            "in": "query",
            "name": "source_branch",
            "required": true,
            "schema": {
              "description": "Source branch name to search for",
              "title": "Source Branch",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CreateMergeRequestResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Vcs-Find-Merge-Request"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Find Merge Request",
        "tags": [
          "external"
        ]
      },
      "post": {
        "description": "Create a merge request with file changes in a VCS repository.\n\nCreates a new branch, applies file operations (create/update/delete),\nand opens a merge request against the target branch.\n\nCallers should use ``GET /merge-requests`` to check for existing MRs\nbefore calling this endpoint to avoid duplicate creation.\n\nArgs:\n    request: Merge request details including repo, auth, and file operations\n\nReturns:\n    CreateMergeRequestResponse with the URL of the created merge request",
        "operationId": "vcs-create-merge-request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMergeRequestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMergeRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Create Merge Request",
        "tags": [
          "external"
        ]
      }
    },
    "/api/v1/external/vcs/repos/owners": {
      "get": {
        "description": "Get OWNERS file data from a Git repository.\n\nFetches OWNERS file approvers and reviewers from GitHub or GitLab repositories.\nResults are cached for performance (TTL configured in settings).\n\nArgs:\n    params: VCSQueryParams with repo_url, owners_file, ref, and secret reference\n\nReturns:\n    RepoOwnersResponse with provider type, approvers, and reviewers lists\n\nRaises:\n    HTTPException:\n        - 500 Internal Server Error: If VCS API call fails or tokens not found\n\nExample:\n    GET /api/v1/external/vcs/repos/owners?url=https://github.com/openshift/osdctl&path=/&ref=master\n    Response:\n    {\n        \"provider\": \"github\",\n        \"approvers\": [\"github_user1\", \"github_user2\"],\n        \"reviewers\": [\"github_user3\"]\n    }",
        "operationId": "vcs-repo-owners",
        "parameters": [
          {
            "description": "Secret Manager URL",
            "in": "query",
            "name": "secret_manager_url",
            "required": true,
            "schema": {
              "description": "Secret Manager URL",
              "title": "Secret Manager Url",
              "type": "string"
            }
          },
          {
            "description": "Path to the secret",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "description": "Path to the secret",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "description": "Specific field within the secret",
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specific field within the secret",
              "title": "Field"
            }
          },
          {
            "description": "Version of the secret",
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Version of the secret",
              "title": "Version"
            }
          },
          {
            "description": "Repository URL (e.g., https://github.com/owner/repo)",
            "in": "query",
            "name": "repo_url",
            "required": true,
            "schema": {
              "description": "Repository URL (e.g., https://github.com/owner/repo)",
              "title": "Repo Url",
              "type": "string"
            }
          },
          {
            "description": "Path to OWNERS file in the repository (e.g., /OWNERS or /path/to/OWNERS)",
            "in": "query",
            "name": "owners_file",
            "required": false,
            "schema": {
              "default": "/OWNERS",
              "description": "Path to OWNERS file in the repository (e.g., /OWNERS or /path/to/OWNERS)",
              "title": "Owners File",
              "type": "string"
            }
          },
          {
            "description": "Git reference (branch, tag, commit SHA)",
            "in": "query",
            "name": "ref",
            "required": false,
            "schema": {
              "default": "master",
              "description": "Git reference (branch, tag, commit SHA)",
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoOwnersResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Repo Owners",
        "tags": [
          "external"
        ]
      }
    },
...
  }
}